# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Talent Bridge API",
    "version": "1.0.0",
    "description": "API documentation for the Talent Bridge platform",
    "contact": {
      "name": "Talent Bridge Team",
      "email": "support@talentbridge.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Development server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from login endpoint"
      }
    },
    "schemas": {
...
      "PaginationInfo": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "description": "Current page number",
            "example": 1
          },
          "limit": {
            "type": "integer",
            "description": "Current page limit",
            "example": 20
          },
          "total": {
            "type": "integer",
            "description": "Total number of items",
            "example": 150
          },
          "totalPages": {
            "type": "integer",
            "description": "Total number of pages",
            "example": 8
          },
          "hasMore": {
            "type": "boolean",
            "description": "Whether there are more results available"
          },
          "nextCursor": {
            "type": "string",
            "description": "Cursor for the next page"
          }
        }
      },
...
      "UserBasic": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "student",
              "school",
              "employer"
            ]
          }
        }
      },
      "ChatMessage": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "Message ID"
          },
          "conversationId": {
            "type": "string",
            "description": "ID of the conversation this message belongs to"
          },
          "sender": {
            "$ref": "#/components/schemas/UserBasic"
          },
          "text": {
            "type": "string",
            "description": "Message text content"
          },
          "attachments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChatAttachment"
            }
          },
          "readBy": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Array of user IDs who have read this message"
          },
          "edited": {
            "type": "boolean",
            "description": "Whether the message has been edited"
          },
          "editedAt": {
            "type": "string",
            "format": "date-time",
            "description": "When the message was last edited"
          },
          "replyTo": {
            "$ref": "#/components/schemas/ChatMessageBasic"
          },
          "type": {
            "type": "string",
            "enum": [
              "text",
              "media",
              "mixed"
            ],
            "description": "Type of message based on content"
          },
          "hasAttachments": {
            "type": "boolean",
            "description": "Whether message has attachments"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/chat/messages": {
      "post": {
        "summary": "Send a message",
        "description": "Send a text message with optional file attachments to a conversation",
        "tags": [
          "Chat"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "conversationId",
                  "content"
                ],
                "properties": {
                  "conversationId": {
                    "type": "string",
                    "description": "ID of the conversation to send message to",
                    "example": "64f456def789abc123456789"
                  },
                  "content": {
                    "type": "string",
                    "description": "Message content/text",
                    "example": "Hello, how are you?"
                  },
                  "attachments": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "description": "Optional file attachments (max 5 files, 10MB each)",
                    "maxItems": 5
                  },
                  "text": {
                    "type": "string"
                  },
                  "replyTo": {
                    "type": "string"
                  }
                }
              }
            },
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "conversationId"
                ],
                "properties": {
                  "conversationId": {
                    "type": "string",
                    "description": "ID of the conversation to send message to",
                    "example": "60f7b3b3b3b3b3b3b3b3b3b3"
                  },
                  "text": {
                    "type": "string",
                    "maxLength": 5000,
                    "description": "Message text content",
                    "example": "Hello, how are you?"
                  },
                  "replyTo": {
                    "type": "string",
                    "description": "ID of message being replied to",
                    "example": "60f7b3b3b3b3b3b3b3b3b3b4"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Message sent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Message sent successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "$ref": "#/components/schemas/ChatMessage"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "201": {
            "description": "Message sent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "object",
                          "properties": {
                            "_id": {
                              "type": "string",
                              "example": "64f789abc123def456789012"
                            },
                            "content": {
                              "type": "string",
                              "example": "Hello, how are you?"
                            },
                            "sender": {
                              "type": "string",
                              "example": "64f123abc456def789012345"
                            },
                            "conversationId": {
                              "type": "string",
                              "example": "64f456def789abc123456789"
                            },
                            "attachments": {
                              "type": "array",
                              "items": {
                                "type": "object",
                                "properties": {
                                  "filename": {
                                    "type": "string"
                                  },
                                  "url": {
                                    "type": "string"
                                  },
                                  "fileType": {
                                    "type": "string"
                                  },
                                  "fileSize": {
                                    "type": "number"
                                  }
                                }
                              }
                            },
                            "timestamp": {
                              "type": "string",
                              "format": "date-time"
                            },
                            "isRead": {
                              "type": "boolean",
                              "example": false
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request - missing required fields or file validation error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Access denied - not a conversation participant"
          },
          "404": {
            "description": "Conversation not found"
          },
          "413": {
            "description": "File too large or too many attachments"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/chat/messages/{conversationId}": {
      "get": {
        "summary": "Get messages from a conversation",
        "description": "Retrieve paginated messages from a specific conversation",
        "tags": [
          "Chat"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "conversationId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Conversation ID",
            "example": "64f456def789abc123456789"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "maximum": 50
            },
            "description": "Page number for pagination"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            },
            "description": "Number of messages per page"
          },
          {
            "in": "query",
            "name": "before",
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "description": "Get messages before this timestamp (for infinite scroll)"
          }
        ],
        "responses": {
          "200": {
            "description": "Messages retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "messages": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "_id": {
                                "type": "string"
                              },
                              "content": {
                                "type": "string"
                              },
                              "sender": {
                                "type": "object",
                                "properties": {
                                  "_id": {
                                    "type": "string"
                                  },
                                  "fullName": {
                                    "type": "string"
                                  },
                                  "avatar": {
                                    "type": "string"
                                  }
                                }
                              },
                              "attachments": {
                                "type": "array",
                                "items": {
                                  "type": "object",
                                  "properties": {
                                    "filename": {
                                      "type": "string"
                                    },
                                    "url": {
                                      "type": "string"
                                    },
                                    "fileType": {
                                      "type": "string"
                                    },
                                    "fileSize": {
                                      "type": "number"
                                    }
                                  }
                                }
                              },
                              "timestamp": {
                                "type": "string",
                                "format": "date-time"
                              },
                              "isRead": {
                                "type": "boolean"
                              },
                              "editedAt": {
                                "type": "string",
                                "format": "date-time",
                                "nullable": true
                              }
                            },
                            "$ref": "#/components/schemas/ChatMessage"
                          }
                        },
                        "pagination": {
                          "type": "object",
                          "properties": {
                            "page": {
                              "type": "integer"
                            },
                            "limit": {
                              "type": "integer"
                            },
                            "total": {
                              "type": "integer"
                            },
                            "pages": {
                              "type": "integer"
                            },
                            "hasMore": {
                              "type": "boolean"
                            }
                          },
                          "$ref": "#/components/schemas/PaginationInfo"
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "Messages retrieved successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid conversation ID or pagination parameters"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Access denied - not a conversation participant"
          },
          "404": {
            "description": "Conversation not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/chat/messages/{conversationId}/read": {
      "patch": {
        "summary": "Mark messages as read",
        "description": "Mark all or specific messages in a conversation as read",
        "tags": [
          "Chat"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "conversationId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Conversation ID",
            "example": "64f456def789abc123456789"
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "messageIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "Specific message IDs to mark as read (optional - if not provided, marks all unread messages)",
                    "example": [
                      "64f789abc123def456789012",
                      "64f890abc123def456789013"
                    ]
                  },
                  "upToMessageId": {
                    "type": "string",
                    "description": "Mark all messages up to this message ID as read",
                    "example": "64f789abc123def456789012"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Messages marked as read successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "markedCount": {
                          "type": "integer",
                          "description": "Number of messages marked as read",
                          "example": 5
                        },
                        "conversationId": {
                          "type": "string"
                        },
                        "unreadCount": {
                          "type": "integer",
                          "description": "New unread count for the user"
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "Messages marked as read successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid conversation ID or message ID"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Access denied - not a conversation participant"
          },
          "404": {
            "description": "Conversation not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  }
}